requirements.txt
click
paho-mqtt >= 2.0.0
python-dotenv